requirements.txt
Django==5.0.3
django-admin-smoke==0.5.0
